+++ packages/react/src/inputs/Select/variants/Combobox/Combobox.test.tsx
@@ -1,24 +1,24 @@
 /* eslint func-names: 0 */
 import { expect } from '@playwright/test'
 
-describe('Dropdown (single Select combobox)', function () {
+describe('Combobox (multiple Select combobox)', function () {
   describe('interactions', function () {
     this.timeout(5e3)
 
     describe('change', function () {
       before(async function () {
         await this.goto(
-          this.storybookUrl('tests-dropdown--dropdown-change-test')
+          this.storybookUrl('tests-combobox--combobox-change-test')
         )
       })
       it('calls onAction on change', async function () {
-        await this.page.locator('button[role=combobox]').click()
+        await this.page.locator('input[role=combobox]').click()
         this.page
-          .locator('text=9887015f-f3ea-49f1-a9ec-808e1fa7b62d')
+          .locator('text=9e7a193d-c6cb-4d53-9a26-e5abc4030e32')
           .click({ noWaitAfter: true })
         const dialog = await this.page.waitForEvent('dialog')
         expect(dialog.message()).toEqual(
-          '5cfc6dda-d0be-4a17-8284-7514817d934c::6eef5d6e-ceaf-418f-8a73-478ead53da1b'
+          '6f72b1df-7e3e-4825-ae1b-200bd136a69f::b7a63920-a6fe-4c46-85fe-aa98b8e9dbd6'
         )
       })
     })